Goodbye WebGL, hello canvas (now a module)
